fix skip empty area

+++ src/main/java/com/rte_france/antares/datamanager_back/repository/model/StStorageEntity.java
@@ -40,7 +40,7 @@ public class StStorageEntity {
     private BigDecimal efficiencyInjection;
 
     @Column(name = "efficiency_withdrawal")
-    private Integer efficiencyWithdrawal;
+    private BigDecimal efficiencyWithdrawal;
 
     @Column(name = "initial_level")
     private BigDecimal initialLevel;
